# Conflicts:
#	ifs-web-service/ifs-competition-mgt-service/src/main/resources/templates/fragments/competition-setup-section-application.html

+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/application/repository/QuestionRepository.java
@@ -19,6 +19,6 @@ public interface QuestionRepository extends PagingAndSortingRepository<Question,
     Question findFirstByCompetitionIdAndSectionIdOrderByPriorityAsc(Long competitionId, Long sectionId);
     List<Question> findByCompetitionIdAndSectionIdOrderByPriorityAsc(Long competitionId, Long sectionId);
     Question findFirstByCompetitionIdAndSectionIdOrderByPriorityDesc(Long competitionId, Long sectionId);
-    List<Question> findByCompetitionIdAndPriorityGreaterThanOrderByPriorityAsc(Long competitionId, Long priority);
+    List<Question> findByCompetitionIdAndPriorityGreaterThanOrderByPriorityAsc(Long competitionId, Integer priority);
     Question findFirstByCompetitionIdAndPriorityGreaterThanOrderByPriorityAsc(Long competitionId, Integer priority);
 }
